get rid of the v1 in the api
